+++ CHANGELOG.md
@@ -339,3 +339,5 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Identify what mod relics, cards, and powers are from (kiooeht)
 * Allowing adding relics to blue pool (kiooeht)
 * Make relic and power onEvokeOrb actually usable (kiooeht)
+* Fix ascension for modded characters (kiooeht)
+* Fix crash if Note For Yourself event (kiooeht)